:recycle: Service 구현체 인터페이스 분리
